+++ libraries/grpc-sdk/src/modules/pushNotifications/types.ts
@@ -2,4 +2,6 @@ export type SendNotificationOptions = {
   body?: string;
   data?: string;
   platform?: string;
+  doNotStore?: boolean;
+  isSilent?: boolean;
 };