+++ modules/communications/src/config/config.ts
@@ -233,6 +233,7 @@ export default {
       phoneNumber: {
         format: 'String',
         default: '',
+        optional: true,
       },
       accountSID: {
         format: 'String',